[refactor] 홈 화면 component 분리
