+++ core/designsystem/src/main/res/values/strings.xml
@@ -28,7 +28,7 @@
     <string name="home_add_interest_festival_button">관심 축제 추가하기</string>
     <string name="home_add_interest_festival_in_item_button">관심 축제로 추가</string>
     <string name="home_add_interest_festival_saved_message">관심 축제로 저장했습니다</string>
-    <string name="home_add_interest_festival_saved_failed_message">관심 축제로 저장했습니다</string>
+    <string name="home_add_interest_festival_saved_failed_message">관심 축제로 저장을 실패했습니다.</string>
     <string name='home_empty_festival_text'>축제 일정 없음</string>
     <string name="home_empty_festival_schedule_description_text">오늘은 축제가 열리는 학교가 없어요</string>
